Made new package gui.
added class Background which makes a panel of a picture from file.

added class LoginPanel which creates a panel with text fields for user to login.

added class Login which creates a layered pane for the full view of the login page.

added class MainView which has main method that creates a frame and shows Login.